producer/requirements.txt
paho-mqtt
apscheduler
protobuf3-to-dict
requests
gtfs-realtime-bindings
